refactor: move share, rename and mapper endpoints to core/api/document.py
Complete the document API consolidation: move update_document_title
(model/rename_doc.py), make_mapped_doc and map_docs (model/mapper.py) and
the share endpoints (share.py) into frappe/core/api/document.py, all
marked @frappe.public(group="Documents") with full annotations and
docstrings.

The share endpoints get proper names in the new module - add_share,
set_share_permission, get_shared_users - since bare add/set_permission/
get_users are too generic for a consolidated module. Business logic
(add_docshare etc.) stays in frappe/share.py.

The old modules keep their dotted paths working via lazy module
__getattr__ aliases: frappe.permissions (which is in the new module's
import chain) imports frappe.share at module level, so plain import
aliases would be circular.
