+++ frappe/model/mapper.py
@@ -9,52 +9,6 @@ from frappe.model.document import Document
 from frappe.utils import cstr
 
 
-@frappe.whitelist()
-def make_mapped_doc(
-	method: str,
-	source_name: str,
-	selected_children: str | list | dict | None = None,
-	args: str | dict | None = None,
-):
-	"""Return the mapped document calling the given mapper method.
-	Set `selected_children` as flags for the `get_mapped_doc` method.
-
-	Called from `open_mapped_doc` from create_new.js"""
-	method = frappe.get_attr(frappe.override_whitelisted_method(method))
-
-	frappe.is_whitelisted(method)
-
-	if selected_children:
-		selected_children = frappe.parse_json(selected_children)
-
-	if args:
-		frappe.flags.args = frappe._dict(frappe.parse_json(args))
-
-	frappe.flags.selected_children = selected_children or None
-
-	return method(source_name)
-
-
-@frappe.whitelist()
-def map_docs(
-	method: str, source_names: str | list, target_doc: Document | dict | str, args: str | dict | None = None
-):
-	"""Return the mapped document calling the given mapper method with each of the given source docs on the target doc.
-
-	:param args: Args as string to pass to the mapper method
-
-	e.g. args: "{ 'supplier': 'XYZ' }"
-	"""
-	method = frappe.get_attr(frappe.override_whitelisted_method(method))
-
-	frappe.is_whitelisted(method)
-
-	for src in frappe.parse_json(source_names):
-		_args = (src, target_doc, frappe.parse_json(args)) if args else (src, target_doc)
-		target_doc = method(*_args)
-	return target_doc
-
-
 def get_mapped_doc(
 	from_doctype,
 	from_docname,
@@ -279,3 +233,8 @@ def map_child_doc(source_d, target_parent, table_map, source_parent=None):
 	target_d.idx = None
 	target_parent.append(target_parentfield, target_d)
 	return target_d
+
+
+# `make_mapped_doc` and `map_docs` moved to frappe.core.api.document; aliases
+# keep the old dotted paths and importers working.
+from frappe.core.api.document import make_mapped_doc, map_docs
